nudity_detect/requirements.txt
tensorflow_gpu==1.13.1
opencv_python
numpy==1.14.5
scipy==1.3.0
matplotlib==3.0.2
Pillow==6.0.0
scikit_learn==0.20.3
torch
torchvision
glob
Pillow